PY175/users_interests/pyproject.toml
[tool.poetry]
name = "users-interests"
version = "0.1.0"
description = ""
authors = ["CovertRob <robby.feconda365@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.3"
pyyaml = "^6.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
